Move fine-tuning and evaluation scripts to examples directory
